refactor the controller folder
