+++ api/src/controllers/user/user.controller.js
@@ -1,11 +1,10 @@
-const userService=require('../services/user.services');
+const userService=require('../../services/user.services');
 
 const getOthersController =async (req,res,next)=>{
     try {
         if(!req.user){
             next(new Error('Your are not authenticated yet get Other people'))
         }
-        console.log(req.user)
         const {user_id}=req.user;
         const others = await userService.getOthers(user_id)
         res.status(200)
