+++ api/src/routers/auth.router.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const authController = require('../controllers/auth.controller');
+const authController = require('../controllers/auth/auth.controller');
 
 router.post('/register',authController.register)
 router.post('/login',authController.login)
