+++ api/src/routers/user.route.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const userControlle = require('../controllers/user.controller');
+const userControlle = require('../controllers/user/user.controller');
 const verifyToken = require('../middlewares/authentication.middleware');
 
 router.get('/get-others',verifyToken,userControlle.getOthersController);
